# apps/api/src/routes/inboundItems.js
const express = require("express");
const { z } = require("zod");
const { getPool } = require("../db");
const { validate } = require("../middleware/validate");
const {
  StockError,
  withTransaction,
  getInboundOrderContext,
  adjustAvailableQty,
  upsertStockTxn,
  softDeleteStockTxn
} = require("../services/stock");
const { syncInboundOrderBillingEvent } = require("../services/billingEvents");
const { getScopedClientId } = require("../middleware/clientScope");

const router = express.Router();

const inboundItemSchema = z.object({
  inbound_order_id: z.coerce.number().int().positive(),
  product_id: z.coerce.number().int().positive(),
  lot_id: z.coerce.number().int().positive().nullable().optional(),
  location_id: z.coerce.number().int().positive().nullable().optional(),
  qty: z.coerce.number().int().positive(),
  invoice_price: z.coerce.number().positive().nullable().optional(),
  currency: z.enum(["KRW", "THB", "USD"]).nullable().optional(),
  remark: z.string().max(500).nullable().optional()
});

function isMysqlForeignKey(error) {
  return error && error.code === "ER_NO_REFERENCED_ROW_2";
}

function isMysqlMissingTable(error) {
  return error && error.code === "ER_NO_SUCH_TABLE";
}

function resolveActorUserId(req, fallbackUserId) {
  const tokenUserId = Number(req.user?.sub || 0);
  if (Number.isFinite(tokenUserId) && tokenUserId > 0) return tokenUserId;
  const fallback = Number(fallbackUserId || 0);
  if (Number.isFinite(fallback) && fallback > 0) return fallback;
  return null;
}

async function appendInboundOrderLog(conn, { inboundOrderId, action, note, actorUserId }) {
  try {
    await conn.query(
      `INSERT INTO inbound_order_logs (inbound_order_id, action, note, actor_user_id)
       VALUES (?, ?, ?, ?)`,
      [inboundOrderId, action, note || null, actorUserId]
    );
  } catch (error) {
    if (!isMysqlMissingTable(error)) {
      throw error;
    }
  }
}

async function validateLotBelongsToProduct(conn, productId, lotId) {
  const [rows] = await conn.query(
    "SELECT id FROM product_lots WHERE id = ? AND product_id = ? AND deleted_at IS NULL",
    [lotId, productId]
  );
  return rows.length > 0;
}

async function resolveLotId(conn, productId, lotId) {
  if (lotId) {
    const validLot = await validateLotBelongsToProduct(conn, productId, lotId);
    if (!validLot) {
      throw new StockError("INVALID_LOT_PRODUCT", "lot_id does not belong to product_id");
    }
    return lotId;
  }

  const [result] = await conn.query(
    `INSERT INTO product_lots (product_id, lot_no, status, deleted_at)
     VALUES (?, 'NO-LOT', 'active', NULL)
     ON DUPLICATE KEY UPDATE deleted_at = NULL, status = 'active', id = LAST_INSERT_ID(id)`,
    [productId]
  );
  return result.insertId;
}

async function getInboundItemWithContext(conn, itemId) {
  const [rows] = await conn.query(
    `SELECT ii.id, ii.inbound_order_id, ii.product_id, ii.lot_id, ii.location_id, ii.qty, ii.invoice_price, ii.currency, ii.remark, ii.created_at, ii.updated_at,
            io.client_id, io.warehouse_id, io.created_by, io.status
     FROM inbound_items ii
     JOIN inbound_orders io ON io.id = ii.inbound_order_id
     WHERE ii.id = ? AND ii.deleted_at IS NULL`,
    [itemId]
  );
  return rows[0] || null;
}

function isInboundReceiptAppliedStatus(status) {
  return status === "received";
}

router.get("/", async (req, res) => {
  const inboundOrderId = req.query.inbound_order_id;

  try {
    const scopedClientId = getScopedClientId(req);
    let query = `SELECT id, inbound_order_id, product_id, lot_id, location_id, qty, invoice_price, currency, remark, created_at, updated_at
                 FROM inbound_items
                 WHERE deleted_at IS NULL`;
    const params = [];

    if (scopedClientId) {
      query += " AND inbound_order_id IN (SELECT id FROM inbound_orders WHERE client_id = ? AND deleted_at IS NULL)";
      params.push(scopedClientId);
    }

    if (inboundOrderId) {
      query += " AND inbound_order_id = ?";
      params.push(inboundOrderId);
    }

    query += " ORDER BY id DESC";

    const [rows] = await getPool().query(query, params);
    res.json({ ok: true, data: rows });
  } catch (error) {
    res.status(500).json({ ok: false, message: error.message });
  }
});

router.get("/:id", async (req, res) => {
  try {
    const scopedClientId = getScopedClientId(req);
    const [rows] = await getPool().query(
      `SELECT id, inbound_order_id, product_id, lot_id, location_id, qty, invoice_price, currency, remark, created_at, updated_at
       FROM inbound_items
       WHERE id = ? AND deleted_at IS NULL
       ${scopedClientId ? "AND inbound_order_id IN (SELECT id FROM inbound_orders WHERE client_id = ? AND deleted_at IS NULL)" : ""}`,
      scopedClientId ? [req.params.id, scopedClientId] : [req.params.id]
    );
    if (rows.length === 0) {
      return res.status(404).json({ ok: false, message: "Inbound item not found" });
    }
    res.json({ ok: true, data: rows[0] });
  } catch (error) {
    res.status(500).json({ ok: false, message: error.message });
  }
});

router.post("/", validate(inboundItemSchema), async (req, res) => {
  const {
    inbound_order_id,
    product_id,
    lot_id,
    location_id = null,
    qty,
    invoice_price = null,
    currency = null,
    remark = null
  } = req.body;

  try {
    const created = await withTransaction(async (conn) => {
      const order = await getInboundOrderContext(conn, inbound_order_id);
      if (!order) {
        throw new StockError("INVALID_ORDER", "Invalid inbound_order_id");
      }
      const resolvedLotId = await resolveLotId(conn, product_id, lot_id);

      const [result] = await conn.query(
        `INSERT INTO inbound_items (inbound_order_id, product_id, lot_id, location_id, qty, invoice_price, currency, remark)
         VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
        [inbound_order_id, product_id, resolvedLotId, location_id, qty, invoice_price, currency, remark]
      );

      if (isInboundReceiptAppliedStatus(order.status)) {
        await adjustAvailableQty(
          conn,
          {
            clientId: order.client_id,
            productId: product_id,
            lotId: resolvedLotId,
            warehouseId: order.warehouse_id,
            locationId: location_id
          },
          qty
        );

        await upsertStockTxn(conn, {
          clientId: order.client_id,
          productId: product_id,
          lotId: resolvedLotId,
          warehouseId: order.warehouse_id,
          locationId: location_id,
          txnType: "inbound_receive",
          qtyIn: qty,
          qtyOut: 0,
          refType: "inbound_item",
          refId: result.insertId,
          createdBy: order.created_by,
          note: remark
        });
      }

      await syncInboundOrderBillingEvent(conn, order.id);
      await appendInboundOrderLog(conn, {
        inboundOrderId: order.id,
        action: "item_create",
        note: `Item added (product=${product_id}, lot=${resolvedLotId}, qty=${qty})`,
        actorUserId: resolveActorUserId(req, order.created_by)
      });

      const [rows] = await conn.query(
        `SELECT id, inbound_order_id, product_id, lot_id, location_id, qty, invoice_price, currency, remark, created_at, updated_at
         FROM inbound_items
         WHERE id = ?`,
        [result.insertId]
      );
      return rows[0];
    });

    return res.status(201).json({ ok: true, data: created });
  } catch (error) {
    if (isMysqlForeignKey(error)) {
      return res.status(400).json({
        ok: false,
        message: "Invalid inbound_order_id, product_id, lot_id or location_id"
      });
    }
    if (error instanceof StockError) {
      return res.status(400).json({ ok: false, code: error.code, message: error.message });
    }
    return res.status(500).json({ ok: false, message: error.message });
  }
});

router.put("/:id", validate(inboundItemSchema), async (req, res) => {
  const {
    inbound_order_id,
    product_id,
    lot_id,
    location_id,
    qty,
    invoice_price,
    currency,
    remark
  } = req.body;

  try {
    const updated = await withTransaction(async (conn) => {
      const prev = await getInboundItemWithContext(conn, req.params.id);
      if (!prev) {
        throw new StockError("NOT_FOUND", "Inbound item not found");
      }

      const nextOrder = await getInboundOrderContext(conn, inbound_order_id);
      if (!nextOrder) {
        throw new StockError("INVALID_ORDER", "Invalid inbound_order_id");
      }
      const resolvedLotId = await resolveLotId(conn, product_id, lot_id);
      const [nextOrderRows] = await conn.query(
        `SELECT status FROM inbound_orders WHERE id = ? AND deleted_at IS NULL LIMIT 1`,
        [inbound_order_id]
      );
      if (nextOrderRows.length === 0) {
        throw new StockError("INVALID_ORDER", "Invalid inbound_order_id");
      }

      const prevApplied = isInboundReceiptAppliedStatus(prev.status);
      const nextApplied = isInboundReceiptAppliedStatus(nextOrderRows[0].status);

      await conn.query(
        `UPDATE inbound_items
         SET inbound_order_id = ?, product_id = ?, lot_id = ?, location_id = ?, qty = ?, invoice_price = ?, currency = ?, remark = ?
         WHERE id = ? AND deleted_at IS NULL`,
        [
          inbound_order_id,
          product_id,
          resolvedLotId,
          location_id || null,
          qty,
          invoice_price || null,
          currency || null,
          remark || null,
          req.params.id
        ]
      );

      if (prevApplied) {
        await adjustAvailableQty(
          conn,
          {
            clientId: prev.client_id,
            productId: prev.product_id,
            lotId: prev.lot_id,
            warehouseId: prev.warehouse_id,
            locationId: prev.location_id
          },
          -Number(prev.qty)
        );
      }

      if (nextApplied) {
        await adjustAvailableQty(
          conn,
          {
            clientId: nextOrder.client_id,
            productId: product_id,
            lotId: resolvedLotId,
            warehouseId: nextOrder.warehouse_id,
            locationId: location_id || null
          },
          qty
        );

        await upsertStockTxn(conn, {
          clientId: nextOrder.client_id,
          productId: product_id,
          lotId: resolvedLotId,
          warehouseId: nextOrder.warehouse_id,
          locationId: location_id || null,
          txnType: "inbound_receive",
          qtyIn: qty,
          qtyOut: 0,
          refType: "inbound_item",
          refId: Number(req.params.id),
          createdBy: nextOrder.created_by,
          note: remark
        });
      } else {
        await softDeleteStockTxn(conn, "inbound_receive", "inbound_item", req.params.id);
      }

      if (Number(prev.inbound_order_id) !== Number(nextOrder.id)) {
        await syncInboundOrderBillingEvent(conn, prev.inbound_order_id);
        await appendInboundOrderLog(conn, {
          inboundOrderId: prev.inbound_order_id,
          action: "item_move_out",
          note: `Item moved out to inbound_order_id=${nextOrder.id} (item=${req.params.id}, qty=${qty})`,
          actorUserId: resolveActorUserId(req, prev.created_by)
        });
      }
      await syncInboundOrderBillingEvent(conn, nextOrder.id);
      await appendInboundOrderLog(conn, {
        inboundOrderId: nextOrder.id,
        action: Number(prev.inbound_order_id) === Number(nextOrder.id) ? "item_update" : "item_move_in",
        note:
          Number(prev.inbound_order_id) === Number(nextOrder.id)
            ? `Item updated (item=${req.params.id}, qty ${prev.qty} -> ${qty})`
            : `Item moved in from inbound_order_id=${prev.inbound_order_id} (item=${req.params.id}, qty=${qty})`,
        actorUserId: resolveActorUserId(req, nextOrder.created_by)
      });

      const [rows] = await conn.query(
        `SELECT id, inbound_order_id, product_id, lot_id, location_id, qty, invoice_price, currency, remark, created_at, updated_at
         FROM inbound_items
         WHERE id = ?`,
        [req.params.id]
      );
      return rows[0];
    });

    return res.json({ ok: true, data: updated });
  } catch (error) {
    if (isMysqlForeignKey(error)) {
      return res.status(400).json({
        ok: false,
        message: "Invalid inbound_order_id, product_id, lot_id or location_id"
      });
    }
    if (error instanceof StockError) {
      const status = error.code === "NOT_FOUND" ? 404 : 400;
      return res.status(status).json({ ok: false, code: error.code, message: error.message });
    }
    return res.status(500).json({ ok: false, message: error.message });
  }
});

router.delete("/:id", async (req, res) => {
  try {
    await withTransaction(async (conn) => {
      const prev = await getInboundItemWithContext(conn, req.params.id);
      if (!prev) {
        throw new StockError("NOT_FOUND", "Inbound item not found");
      }

      if (isInboundReceiptAppliedStatus(prev.status)) {
        await adjustAvailableQty(
          conn,
          {
            clientId: prev.client_id,
            productId: prev.product_id,
            lotId: prev.lot_id,
            warehouseId: prev.warehouse_id,
            locationId: prev.location_id
          },
          -Number(prev.qty)
        );
      }

      await conn.query(
        "UPDATE inbound_items SET deleted_at = NOW() WHERE id = ? AND deleted_at IS NULL",
        [req.params.id]
      );

      await softDeleteStockTxn(conn, "inbound_receive", "inbound_item", req.params.id);
      await syncInboundOrderBillingEvent(conn, prev.inbound_order_id);
      await appendInboundOrderLog(conn, {
        inboundOrderId: prev.inbound_order_id,
        action: "item_delete",
        note: `Item deleted (item=${req.params.id}, product=${prev.product_id}, lot=${prev.lot_id}, qty=${prev.qty})`,
        actorUserId: resolveActorUserId(req, prev.created_by)
      });
    });

    return res.json({ ok: true });
  } catch (error) {
    if (error instanceof StockError) {
      const status = error.code === "NOT_FOUND" ? 404 : 400;
      return res.status(status).json({ ok: false, code: error.code, message: error.message });
    }
    return res.status(500).json({ ok: false, message: error.message });
  }
});

module.exports = router;